requirements.txt
langchain
langchain-community
langgraph
openai
oracledb
psycopg2-binary
python-dotenv
